Refactor routes to use headers for uid
It seems a better method to use a header with every request in order
to retrieve the correct plants. Also added a bit more verification that
the uid is in fact a string, as well as adding uid verification to
the delete plant route, otherwise no id would be needed to delete any
plant.

+++ server/src/controllers/plants.ts
@@ -2,12 +2,10 @@ import express from 'express';
 export const plantsRouter = express.Router();
 import { sql } from '../utils/db';
 
-interface GetBody {
-	uid: string;
-}
-
 plantsRouter.get('/', async (req, res) => {
-	const { uid } = req.body as GetBody;
+	const uid = req.get('uid');
+	if (typeof uid !== 'string') throw new Error('Uid is not a string!');
+
 	const plants = await sql`
     SELECT plants.id, name, schedule, array_agg(water.date ORDER BY water.date ASC) as watered,
     MAX(water.date) + schedule as next_water
@@ -23,18 +21,20 @@ plantsRouter.get('/', async (req, res) => {
 interface PostBody {
 	name: string;
 	schedule: number;
-	uid: string;
 }
 
 // Create new plant
 plantsRouter.post('/', async (req, res) => {
-	const { name, schedule, uid } = req.body as PostBody;
+	const { name, schedule } = req.body as PostBody;
+	const uid = req.get('uid');
+	if (typeof uid !== 'string') throw new Error('Uid is not a string!');
+
 	const plants = await sql`
     INSERT INTO plants 
         (name, schedule, uid) 
     VALUES 
         (${name}, ${schedule}, ${uid})
-    RETURNING name, schedule, uid
+    RETURNING name, schedule
     `;
 	return res.status(201).json(plants);
 });
@@ -43,7 +43,7 @@ interface BodyWater {
 	plant_id: number;
 	date: Date;
 }
-
+// Post single date
 plantsRouter.post('/water', async (req, res) => {
 	const { plant_id, date } = req.body as BodyWater;
 	const plant = await sql`
@@ -74,6 +74,9 @@ plantsRouter.delete('/water', async (req, res) => {
 // Delete one plant and associated dates
 plantsRouter.delete('/', async (req, res) => {
 	const { plant_id } = req.body as BodyWater;
+	const uid = req.get('uid');
+	if (typeof uid !== 'string') throw new Error('Uid is not a string!');
+
 	await sql`
     DELETE FROM water 
     WHERE 
@@ -83,6 +86,8 @@ plantsRouter.delete('/', async (req, res) => {
     DELETE FROM plants
     WHERE 
         id = ${plant_id}
+    AND 
+        uid = ${uid}
     RETURNING id, name
     `;
 	return res.status(200).json(deletedPlant);
